+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## v0.59.0
 
-- improved HTML parser logic again, specifically in the case of how it collects contents of a document;
-- fallbakc to the English language for the stop words in cases when language detection is not reliable.
+- use different segmentation logic based on the `github.com/go-ego/gse` segmenter for Chinese & Japanese languages;
+- improved HTML parser logic: optimised the way it collects contents of a document and improved logic for splitting into sentences;
+- fallback to the English language for the stop words in cases when language detection is not reliable;
+- added `lang` option to the CLI to be able to provide the language of the document;
+- bumped `github.com/zoomio/stopwords` to `0.11.0`.
 
 ## v0.58.0